[Fix] 도메인 UI분리 롤백
